osx/du.md: fix depth option, plus minor tweaks
The version I have uses `-d` for the depth option, not `-depth`. Here's the version string as shown by `man du`:
```
BSD                              June 2, 2004                              BSD
```

In addition to this fix, this commit changes the tokens to clarify that they are paths, and changes "directory" to "folder", which is shorter and more accessible.